+++ lib/fluid_animations.dart
@@ -1,8 +1,10 @@
-/// A Very Good Project created by Very Good CLI.
+/// A library for creating smooth and responsive animations 🔥
 library;
 
+export 'src/curve.dart';
 export 'src/fluid_spring.dart';
 export 'src/fluid_transition_builder.dart';
+export 'src/keyframe.dart';
 export 'src/simulation.dart';
 export 'src/tweens.dart';
 export 'src/velocity_helper.dart';