+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog],
-and this project adheres to [Semantic Versioning].
+## 1.1.0 - 2024-09-24
 
+- Added `KeyframeAnimation` for complex animations.
+- Added `SpringCurve` to convert spring to curve.
+- Added `SpringTween`
 
 ## 1.0.0 - 2024-05-12